pyproject.toml
[project]
name = "infotransform"
version = "0.1.0"
description = "Transform any file type into structured, actionable data using AI-powered analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "markitdown[all]>=0.1.2",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.1",
    "aiosqlite>=0.20.0",
    "jinja2>=3.1.2",
    "pydantic-ai>=0.0.9",
    "pandas>=2.0.0",
    "aiohttp>=3.12.13",
    "openpyxl>=3.1.5",
    "tiktoken>=0.9.0",
    "ruff>=0.12.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.25.0",
]

[dependency-groups]
dev = [
    "pip-licenses>=5.5.0",
]
